// app/common/enum.ts
export enum FILTER_NFT_TYPE_GALIX_MARKET {
  HERO = "Hero",
  RESOURCE = "Resource",
  LAND = "Land",
  MYSTERY_BOX = "Box",
  PACKAGE = "Package",
}
// ...
export enum SERVICE_ID {
  _9DNFT = "10000",
  _GALIXCITY = "20000",
  _SOUL_REALM = "30000",
  _HEAVEN_SWORD = "40000",
  _MARSWAR = "50000",
  _NARUTO = "60000",
  _FLASHPOINT = "70000",
  _RICHWORK_FARM_FAMILY = "80000",
  _OPS_ALPHA = "11000",
}