// app/modules/graphql/types/generated.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
export type Exact<T extends Record<string, unknown>> = {
  [K in keyof T]: T[K];
};
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & {
  [SubKey in K]?: Maybe<T[SubKey]>;
};
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & {
  [SubKey in K]: Maybe<T[SubKey]>;
};
/** All built-in and custom scalars, mapped to their actual values */
export interface Scalars {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  BigDecimal: number;
  BigInt: number;
  Long: number;
}

export interface AskOrder {
  __typename?: "AskOrder";
  /** Collection */
  collection: Collection;
  /** Block hash */
  id: Scalars["ID"];
  /** isTradable */
  isTradable: Scalars["Boolean"];
  /** NFT */
  nft: Nft;
  /** Ask Price */
  price: Scalars["BigDecimal"];
  /** Seller */
  seller: User;
  /** Updated at */
  updatedAt: Scalars["BigInt"];
}

export interface BidOrder {
  __typename?: "BidOrder";
  /** Buyer */
  buyer: User;
  /** Collection */
  collection: Collection;
  /** ask ID + buyer ID */
  id: Scalars["ID"];
  /** isTradable */
  isTradable: Scalars["Boolean"];
  /** NFT */
  nft: Nft;
  /** Bid Price */
  price: Scalars["BigDecimal"];
  /** Updated at */
  updatedAt: Scalars["BigInt"];
}

export interface Collection {
  __typename?: "Collection";
  /** ID (address) */
  id: Scalars["ID"];
  /** Name */
  name: Scalars["String"];
  /** List of NFTs */
  nfts: Nft[];
  /** Symbol */
  symbol: Scalars["String"];
}

export interface Metadata {
  __typename?: "Metadata";
  /** List of attribute */
  attributes: MetadataAttribute[];
  /** Description */
  description: Scalars["String"];
  /** ID (cid) */
  id?: Scalars["ID"];
  /** Image */
  image: Scalars["String"];
  /** Name */
  name: Scalars["String"];
  /** NFT */
  nft?: Maybe<Nft>;
}

export interface MetadataAttribute {
  __typename?: "MetadataAttribute";
  /** display_type */
  display_type?: Scalars["String"];
  /** Metadata + incr */
  id?: Scalars["ID"];
  /** Metadata (cid) */
  metadata?: Metadata;
  /** trait_type */
  trait_type?: Scalars["String"];
  /** value */
  value?: Scalars["String"];
}

export interface Nft {
  __typename?: "NFT";
  /** current ask ID */
  askId?: Maybe<Scalars["String"]>;
  /** ask Price */
  askPrice: Scalars["BigDecimal"];
  /** asks */
  asks: AskOrder[];
  /** current bid ID */
  bidId?: Maybe<Array<Scalars["String"]>>;
  /** bids */
  bids: BidOrder[];
  /** Collection address */
  collection: Collection;
  /** CollectionAddress + tokenId */
  id: Scalars["ID"];
  /** isTradable */
  isTradable: Scalars["Boolean"];
  /** Metadata cid */
  metadata: Metadata;
  /** Metadata Type */
  metadataLevel: Scalars["Int"];
  /** Metadata Name */
  metadataNames: Scalars["String"];
  /** Metadata Type */
  metadataType: Scalars["String"];
  /** metadata NftType Type */
  metadataNftType: Scalars["String"];
  /** metadata Rarity */
  metadataRarity: Scalars["BigInt"];
  /** metadata Quality */
  metadataQuality: Scalars["BigInt"];
  /** metadata Index */
  metadataIndex: Scalars["BigInt"];
  /** Owner */
  owner: User;
  /** seller */
  seller: User;
  /** tokenId */
  tokenId: Scalars["BigInt"];
  /** Transaction history */
  transactionHistory: Transaction[];
  /** Updated at */
  updatedAt: Scalars["BigInt"];
}

export interface Transaction {
  __typename?: "Transaction";
  block: Scalars["BigInt"];
  /** Collection */
  collection: Collection;
  id: Scalars["ID"];
  /** Kind */
  kind: TransactionKind;
  /** NFT */
  nft: Nft;
  /** price (in COGI) */
  price: Scalars["BigDecimal"];
  /** transfer Recipient */
  recipient: User;
  /** transfer Sender */
  sender: User;
  /** Timestamp */
  timestamp: Scalars["BigInt"];
}

export enum TransactionKind {
  Bid = "Bid",
  Burn = "Burn",
  Cancel = "Cancel",
  Ino = "INO",
  List = "List",
  Mint = "Mint",
  Sale = "Sale",
  Transfer = "Transfer",
}

export interface User {
  __typename?: "User";
  /** List of NFTs */
  askNfts: Nft[];
  /** User address */
  id: Scalars["ID"];
  /** NFTs */
  nfts: Nft[];
  /** Number of tokens */
  numberTokens: Scalars["BigInt"];
  /** Number of tokens currently listed */
  numberTokensListed: Scalars["BigInt"];
  /** Number of tokens purchased */
  numberTokensPurchased: Scalars["BigInt"];
  /** Number of tokens sold */
  numberTokensSold: Scalars["BigInt"];
}